requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
